requirements.txt
annotated-types==0.7.0
attrs==25.4.0
beautifulsoup4==4.14.2
bottle==0.13.4
bs4==0.0.2
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
customtkinter==5.2.2
darkdetect==0.8.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httplib2==0.31.0
idna==3.11
outcome==1.3.0.post0
packaging==25.0
proto-plus==1.26.1
protobuf==5.29.5
proxy_tools==0.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.29.0
pydantic==2.12.4
pydantic_core==2.41.5
PyGObject==3.54.5
pyparsing==3.2.5
PySide6==6.10.1
PySide6_Addons==6.10.1
PySide6_Essentials==6.10.1
PySocks==1.7.1
python-dotenv==1.2.1
pywebview==6.1
QtPy==2.4.3
requests==2.32.5
rsa==4.9.1
selenium==4.38.0
shiboken6==6.10.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
tqdm==4.67.1
trio==0.32.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
webdriver-manager==4.0.2
websocket-client==1.9.0
wsproto==1.3.1
